Make "boot=" parameter mandatory
When running in initramfs, no file system is automatically mounted by
the kernel, so we need to mount the fs containing the rootfs image.
If the "boot=" parameter is missing, we cannot do that, so it's simpler
to give up immediately than getting stranded later.

In theory mininit could be used for non-loop boot sequences and not
require "boot=" in that case. But those scenarios can be handled by
just kernel configuration, so there is no need to have mininit support
them.