When the path passed to losetup is too long, return an error
Previously the code would truncate the path and try to continue with
the truncated version, which is a bad idea: most likely it will lead
to a misleading file-not-found error later on, but if the operation
would somehow succeed then it is working on the wrong path.

If we can't give the user what they asked for, returning an error is
the only right thing to do. I picked ENAMETOOLONG rather than ERANGE
since the exceeded limit is hardwired in the OS.