Hardcode rootfs file name as "rootfs.squashfs"
The advantage is that we need one fewer command line argument.
All the other file names on the system partition were already
hardcoded, so we won't lose much flexibility because of this.

I think the rootfs name argument dates back to the time when mininit
did not play a role in updates: the original Dingux had to be updated
offline, by writing the new files to the SD card from a PC.