Removed file system autodetection
Since we hardcoded the rootfs and system partition file system types,
the autodetection code wasn't used anymore.

It wasn't real autodetection by looking for signatures, it just tried
mounting as each of the file systems listed in /proc/filesystems until
one succeeds. While this works, it is not very efficient and it can
also lead to a lot of logging about failed mounts.

This also brings us closer to not having to mount /proc.