Removed "init=" argument
One reason for removing it is that I want to be able to run mininit
from the system partition instead of from initramfs. In that case,
mininit is the actual init process and the kernel "init=" argument
will be used to point to mininit.

Another reason for removing it is that the kernel passes all arguments
it doesn't know to the init process. But "init=" is a kernel argument,
so it is not passed in this way and can only be retrieved from procfs.