Removed mounting of /proc
The parsing of /proc/cmdline is no longer necessary since we now use
argv and the environment instead. The parsing of /proc/filesystems
was removed earlier. That leaves nothing using /proc anymore, so there
is no longer a neeed to mount it.

Also the helper functions to read and parse files from /proc were
removed.