Only mount /boot read-write if we have an update to perform
Mount it read-only first, then check for updates and remount read-write
only if needed. Also remount as read-only as soon as possible.

Previously there was no sync if only the modules image was upgraded.
Probably the remount as read-only will sync, but I haven't verified
that, so we do an explicit sync() call just to be sure.

Update failures are no longer considered fatal: we can still try to
boot the older version.