Use O_PATH to indicate we want to use the fd for unlinkat() only
The O_PATH flag tells open() that we only need a path reference, not
a file descriptor that we can do I/O through. That should make the
operation a lot cheaper to perform.

Also remove third argument to open(): since we are not creating a file,
this argument has no effect.