Write log messages to kernel log
This makes them accessible after boot and it provides time stamps that
can be useful for rudimentary profiling. Note that the kernel log is
also echoed on the console during early boot, so the only thing we lose
is the coloring of the log messages.

The syslog priority markers could be computed, but not without runtime
overhead or very nasty build-time hacks, so I decided to hardcode them.

I also tweaked the messages themselves a bit.