Support a system partition that is initially mounted read-write
In mininit-syspart, if the system partition was initially mounted
read-write and no update was performed, it would still be read-write
when starting the rootfs init. For consistency with mininit-initramfs,
we will remount it read-only.

If the system partition was initially mounted read-write, the remount
before updates can be skipped.

Elevate the update log messages from DEBUG to INFO level, so they will
show up in the log by default.