Use strncpy instead of strlcpy to copy splashkill process name
In this particular use, the fact that strncpy doesn't guarantee
a terminating '\0' when the string only just fits is actually an
advantage, since the original string already has a '\0' that is
not included in the passed size.

Also, strlcpy is not available in glibc. It is available in uClibc
and musl though.